refactor(workflow): extract Block base into block_base.py (stack foundation)
Introduce models/block_base.py holding Block + shared helpers (capture_block_download_baseline,
warn_if_file_download_max_steps_low, jinja_sandbox_env, CURRENT_DATE_FORMAT,
MAX_STEPS_DOWNLOAD_WARNING_THRESHOLD). block.py re-exports them for back-compat. This is the
foundational base module the rest of the block.py split builds on; subsequent submodules import
Block from block_base to break the import cycle.

block_base.py taken verbatim from SKY-11658 (@855072ac4).
